# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Collective API",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082/api"
    }
  ],
  "tags": [
    {
      "name": "Posts"
    },
    {
      "name": "Comments",
      "description": "User comments on posts."
    },
    {
      "name": "Todos",
      "description": "Todo list"
    }
  ],
  "paths": {
    "/comments": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get all comments.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "Add a comment to a post.",
        "requestBody": {
          "description": "The comment to add. Note: {id} will be ignored in the request body.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Comment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success. The newly created comment will be returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        }
      }
    },
    "/comments/{commentId}": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get a comment.",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "Id of comment to retrieve.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Comments"
        ],
        "summary": "Delete a comment.",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "Id of the comment to delete.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete was successful."
          }
        }
      }
    },
    "/posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get all posts.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "Add a post.",
        "requestBody": {
          "description": "The post to add. Note: {id} will be ignored in the request body.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success. The newly created post will be returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{postId}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get a post.",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "Id of post to retrieve.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "summary": "Delete a post.",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "Id of the post to delete.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete was successful."
          }
        }
      }
    },
    "/posts/votes/increase/{postId}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Upvote a post.",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "Id of the post to upvote.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. The upvoted post will be returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
    "/posts/votes/decrease/{postId}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Downvote a post.",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "Id of the post to downvote.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success. The downvoted post will be returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
    "/todos": {
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "Get all todos.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedTodo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Todos"
        ],
        "summary": "Add a todo.",
        "requestBody": {
          "description": "The todo to add.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Todo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success. The newly created todo will be returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedTodo"
                }
              }
            }
          }
        }
      }
    },
    "/todos/{todoId}": {
      "get": {
        "tags": [
          "Todos"
        ],
        "summary": "Get a todo.",
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "description": "Id of todo to retrive.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedTodo"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Todos"
        ],
        "summary": "Update a todo.",
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "description": "Id of todo to update.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "The todo updates.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Todo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success. The updated todo will be returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedTodo"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Todos"
        ],
        "summary": "Delete a todo.",
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "description": "Id of todo to delete.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete was successful."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "string"
          },
          "post_id": {
            "type": "integer",
            "format": "int32",
            "description": "The id of the post that this comment belongs to."
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "author": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "votes": {
            "type": "integer",
            "format": "int32"
          },
          "img_url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "Todo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "complete": {
            "type": "boolean"
          }
        }
      },
      "SavedTodo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "complete": {
            "type": "boolean"
          }
        }
      }
    }
  }
}